[SDK-223] Simplify Ultracite commands and regression support
