+++ .github/workflows/detect-breaking-changes.yml
@@ -48,7 +48,5 @@ jobs:
 
       - name: Detect breaking changes
         run: |
-          # Try to check out previous versions of the breaking change detection script. This ensures that
-          # we still detect breaking changes when entire files and their tests are removed.
-          git checkout "$BASE_SHA" -- ./scripts/detect-breaking-changes 2>/dev/null || true
-          DETECT_BREAKING_CHANGES_COMPAT=1 ./scripts/detect-breaking-changes "$BASE_SHA"
+          # The detector discovers tests from the base revision, including removed resources.
+          ./scripts/detect-breaking-changes "$BASE_SHA"
