--- scripts/format
@@ -1,12 +1,0 @@
-#!/usr/bin/env bash
-
-set -e
-
-cd "$(dirname "$0")/.."
-export PATH="$PWD/node_modules/.bin:$PATH"
-
-echo "==> Running Ultracite fixes"
-./node_modules/.bin/ultracite fix .
-
-echo "==> Running oxfmt --write"
-./node_modules/.bin/oxfmt --write .
