--- scripts/lint
@@ -1,48 +1,0 @@
-#!/usr/bin/env bash
-
-set -e
-
-cd "$(dirname "$0")/.."
-export PATH="$PWD/node_modules/.bin:$PATH"
-
-if [ "${DETECT_BREAKING_CHANGES_COMPAT:-}" = "1" ]; then
-  echo "==> Preparing historical breaking-change tests"
-  # Historical tests are checked against the current SDK and tsconfig. The
-  # embeddings tests on main briefly used Array.prototype.at(), but the SDK still
-  # targets older JS libs, so normalize that syntax before the compatibility lint.
-  if [ -e tests/api-resources/embeddings.test.ts ]; then
-    node <<'EOF'
-const fs = require('fs');
-
-const path = 'tests/api-resources/embeddings.test.ts';
-let contents = fs.readFileSync(path, 'utf8');
-contents = contents
-  .replaceAll('response.data?.at(0)?.embedding.at(0)', 'response.data?.[0]?.embedding[0]')
-  .replaceAll('response.data?.at(0)?.embedding', 'response.data?.[0]?.embedding');
-fs.writeFileSync(path, contents);
-EOF
-  fi
-
-  # Old test files are checked for compatibility with the current SDK, not
-  # formatting drift against the current Oxfmt version.
-  find tests -name '*.test.ts' -print0 | xargs -0 ./node_modules/.bin/oxfmt --write
-fi
-
-echo "==> Running Ultracite checks"
-./node_modules/.bin/ultracite check .
-
-echo "==> Running Oxlint regressions"
-node --test scripts/oxlint-regression.test.cjs
-
-echo "==> Building"
-./scripts/build
-
-echo "==> Checking types"
-./node_modules/typescript/bin/tsc
-
-echo "==> Running Are The Types Wrong?"
-./node_modules/.bin/attw --pack dist -f json >.attw.json || true
-node scripts/utils/attw-report.cjs
-
-echo "==> Running publint"
-./node_modules/.bin/publint dist
